Add "Audience Measurement" use case to the HitList
I've tried to capture the Audience Measurement example that Sanjay, Kevin, Wendall, Charlie, and others discussed in yesterday's call.

There was a lot of complexity here, in part because we went into more depth discussing real-world use cases.  Let's leave this as an open PR for a bit, and give the contributors to the discussion a chance to improve on my summary before we commit.