Update OnlineAdvertisingCapabilitiesHitList.md
We should reference the admetrics document here. 
I also think that consent will have to be separately obtained for the categorisation of an individual by an audience provider, we should create a paragraph on that. It might be possible to get a nod from regulators for "simple" web audience measurement because an exception was proposed for that in the European Parliament's agreed draft of the ePrivacy Regulation (by the regulators), but getting acceptance for an external audience provider will be problematic, especially if the categorisation involved "sensitive" personal data e.g. political opinion, sexual orientation, income bracket etc.