Add testing opportunities for various proposals.
As some of these proposals are reaching more mature levels they are getting to the point where they may be available for testing and feedback to authors. I would like to add a section to make these opportunities clear and have a central place in which to find them. This will help generate additional discussion and feedback on these proposals from this group. I may have left some out, this is not intentional. Future testing opportunities can be added via PR.